refactor(web): split page components by route
