+++ crates/site/web/build.rs
@@ -2,6 +2,7 @@ fn main() {
     println!("cargo:rerun-if-changed=src");
     println!("cargo:rerun-if-changed=style");
 
+    // `assets::STYLESHEET` registers the stylesheet rendered by this build integration.
     topcoat::tailwind::BuildConfig::new()
         .input("crates/site/web/style/tailwind.css")
         .cwd("../../..")
